Corrige o espelho DUnitX (AreEqual Int64) e o GUID do DFeSimulador.dproj
Achados pela primeira compilacao no Delphi: Assert.AreEqual(3, MaxNSU) nao infere
o tipo generico (Integer x Int64; usar Int64(3)), e o DFeSimulador.dproj herdou o
ProjectGuid do SpikeHorse (que veio do sample do Horse) -- a IDE gerou um novo.
O .dproj entra como a IDE o reescreveu, como os demais do repo.

+++ tests/Unit/DFe.SimuladorServidorTests.pas
@@ -272,7 +272,7 @@ begin
   CarregarCenario(FCaminho, FSim);
   LResp := FSim.Consultar('11222333000181', 'RS', 0);
   Assert.AreEqual(1, Integer(Length(LResp.Lote.Itens)));
-  Assert.AreEqual(3, LResp.Lote.MaxNSU);
+  Assert.AreEqual(Int64(3), LResp.Lote.MaxNSU);
 end;
 
 procedure TDFeSimuladorCenarioTests.Falhas_SaoEnfileiradasNaOrdem;
